test(cli): reorganize e2e tests and add more tests for cli commands
