+++ cli/package.json
@@ -14,10 +14,17 @@
     "watch": "tsc --watch",
     "build": "tsc",
     "postbuild": "cp package.json ./build",
-    "test": "nyc ts-mocha --exit tests/*.test.ts",
+    "test": "nyc ts-mocha --exit tests/**/*.test.ts",
     "test:e2e": "ts-mocha --exit tests/e2e.test.ts",
     "test:e2e-subsidy": "ts-mocha --exit tests/e2e.subsidy.test.ts",
-    "test:keyChange": "ts-mocha --exit tests/keyChange.test.ts"
+    "test:keyChange": "ts-mocha --exit tests/keyChange.test.ts",
+    "test:unit": "nyc ts-mocha --exit tests/unit/*.test.ts",
+    "test:airdrop": "nyc ts-mocha --exit tests/unit/airdrop.test.ts",
+    "test:genPubKey": "ts-mocha --exit tests/unit/genPubKey.test.ts",
+    "test:genKeypair": "ts-mocha --exit tests/unit/genKeyPair.test.ts",
+    "test:timeTravel": "ts-mocha --exit tests/unit/timeTravel.test.ts",
+    "test:fundWallet": "ts-mocha --exit tests/unit/fundWallet.test.ts",
+    "test:utils": "ts-mocha --exit tests/unit/utils.test.ts"
   },
   "dependencies": {
     "@commander-js/extra-typings": "^11.1.0",
@@ -57,10 +64,11 @@
     ],
     "all": true,
     "exclude": [
-      "**/tests/*.ts",
       "**/*.js",
       "**/*.d.ts",
-      "hardhat.config.ts"
+      "hardhat.config.ts",
+      "tests/**/*.ts",
+      "ts/index.ts"
     ],
     "branches": ">50%",
     "lines": ">50%",
