+++ cli/tests/unit/genPubkey.test.ts
@@ -1,10 +1,11 @@
 import { expect } from "chai";
-import { genKeyPair, genMaciPubKey } from "maci-cli";
 import { genPubKey } from "maci-crypto";
 import { PubKey, PrivKey } from "maci-domainobjs";
 
+import { genKeyPair, genMaciPubKey } from "../../ts";
+
 describe("genMaciPubkey CLI subcommand", () => {
-  it("genMaciPubkey should output a correct public key", () => {
+  it("should output a valid public key", () => {
     const keypair = genKeyPair(true);
     const pubKey = genMaciPubKey(keypair.privateKey, true);
 
@@ -15,5 +16,9 @@ describe("genMaciPubkey CLI subcommand", () => {
     const unserializedPk = PubKey.deserialize(keypair.publicKey);
     expect(unserializedPk.rawPubKey[0].toString()).to.eq(pk2[0].toString());
     expect(unserializedPk.rawPubKey[1].toString()).to.eq(pk2[1].toString());
+  });
+
+  it("should throw when given an invalid private key", () => {
+    expect(() => genMaciPubKey("invalid", true)).to.throw("Invalid private key");
   });
 });
